requirements.txt
azure-mgmt-resource>=1.1.0
azure-mgmt-compute>=2.1.0
azure-mgmt-network>=1.4.0
azure-mgmt-keyvault>=0.40.0
azure-keyvault>=0.3.7
haikunator